refactor: modify product model

+++ src/models/productModel.ts
@@ -30,7 +30,7 @@ export function productFactory(sequelize): typeof Product {
                 allowNull: false,
             },
             price: {
-                type: DataTypes.INTEGER,
+                type: DataTypes.FLOAT,
                 allowNull: false,
             },
             category: {
